Added TestBitcoinTill
Added TestWalletKitThread

Both required a bit of rewriting of their target classes.

I have so many wallets in the bitcoin_test_files/ directory as tests on the same WalletKitThread tend to trip over each other and cause errors that are not really within the purview of our project as it stands. I am prepared to change this strategy if everyone else finds it incredibly irritating.

I dabbled with testing Exceptions with JMockit, but found it unintelligible. As such, most of the missing line coverage is inside Exception catches. Maybe we should discuss whether this is an acceptable state of affairs in Monday's meeting.